Added and delete it
This first lesson was a little complicated, since the lessons they are not really good at explination, it gets hard to follow specially when sometimes they tell you to delete stuff and the next steps they don't show what is going to look once you finish adding the election.db and how it behaves.  Luckly, I had help from a fellow college that had a grasp of what they are talkingf about but she has other issues to be solved.